separated some functions in systems into utils.rs and tile_makers.rs
