+++ src/systems.rs
@@ -2,7 +2,9 @@ use crate::{
     app::LdtkEntityMap,
     assets::{LdtkAsset, LdtkExternalLevel, TilesetMap},
     components::*,
-    ldtk::{EntityDefinition, TileInstance, TilesetDefinition, Type},
+    ldtk::{EntityDefinition, TilesetDefinition, Type},
+    tile_makers::*,
+    utils::*,
 };
 
 use bevy::prelude::*;
@@ -408,184 +410,4 @@ fn spawn_level(
             }
         }
     }
-}
-
-fn int_grid_index_to_tile_pos(
-    index: usize,
-    layer_width_in_tiles: i32,
-    layer_height_in_tiles: i32,
-) -> TilePos {
-    let tile_x = index as u32 % layer_width_in_tiles as u32;
-    let tile_y =
-        layer_height_in_tiles as u32 - ((index as u32 - tile_x) / layer_width_in_tiles as u32) - 1;
-
-    TilePos(tile_x, tile_y)
-}
-
-fn tile_pos_to_invisible_tile(_: TilePos) -> Option<Tile> {
-    Some(Tile {
-        visible: false,
-        ..Default::default()
-    })
-}
-
-fn tile_pos_to_tile_maker(
-    layer_height_in_tiles: i32,
-    tileset_definition: &TilesetDefinition,
-    grid_tiles: Vec<TileInstance>,
-) -> impl FnMut(TilePos) -> Option<Tile> {
-    let tile_grid_size = tileset_definition.tile_grid_size;
-    let tileset_width_in_tiles = tileset_definition.c_wid;
-
-    let grid_tile_map: HashMap<TilePos, TileInstance> = grid_tiles
-        .into_iter()
-        .map(|t| {
-            (
-                TilePos(
-                    (t.px[0] / tileset_definition.tile_grid_size) as u32,
-                    layer_height_in_tiles as u32
-                        - (t.px[1] / tileset_definition.tile_grid_size) as u32
-                        - 1,
-                ),
-                t,
-            )
-        })
-        .collect();
-
-    move |tile_pos: TilePos| -> Option<Tile> {
-        match grid_tile_map.get(&tile_pos) {
-            Some(tile_instance) => {
-                let tileset_x = tile_instance.src[0] / tile_grid_size;
-                let tileset_y = tile_instance.src[1] / tile_grid_size;
-                let (flip_x, flip_y) = match tile_instance.f {
-                    1 => (true, false),
-                    2 => (false, true),
-                    3 => (true, true),
-                    _ => (false, false),
-                };
-                Some(Tile {
-                    texture_index: (tileset_y * tileset_width_in_tiles + tileset_x) as u16,
-                    flip_x,
-                    flip_y,
-                    ..Default::default()
-                })
-            }
-            None => None,
-        }
-    }
-}
-
-fn tile_pos_to_tile_bundle_if_int_grid_nonzero_maker(
-    mut tile_maker: impl FnMut(TilePos) -> Option<Tile>,
-    int_grid_csv: &Vec<i32>,
-    layer_width_in_tiles: i32,
-    layer_height_in_tiles: i32,
-) -> impl FnMut(TilePos) -> Option<TileBundle> {
-    let nonzero_map: HashMap<TilePos, bool> = int_grid_csv
-        .iter()
-        .enumerate()
-        .map(|(i, v)| {
-            (
-                int_grid_index_to_tile_pos(i, layer_width_in_tiles, layer_height_in_tiles),
-                *v == 0,
-            )
-        })
-        .collect();
-    move |tile_pos: TilePos| -> Option<TileBundle> {
-        match nonzero_map.get(&tile_pos) {
-            Some(nonzero) if *nonzero => tile_maker(tile_pos).map(|tile| TileBundle {
-                tile,
-                ..Default::default()
-            }),
-            _ => None,
-        }
-    }
-}
-
-fn tile_pos_to_tile_bundle_maker(
-    mut tile_maker: impl FnMut(TilePos) -> Option<Tile>,
-) -> impl FnMut(TilePos) -> Option<TileBundle> {
-    move |tile_pos: TilePos| -> Option<TileBundle> {
-        match tile_maker(tile_pos) {
-            Some(tile) => Some(TileBundle {
-                tile,
-                ..Default::default()
-            }),
-            None => None,
-        }
-    }
-}
-
-/// Similar to LayerBuilder::new_batch, except it doesn't consume the LayerBuilder
-///
-/// This allows for more methods to be performed on the LayerBuilder before building it.
-/// However, the performance cons of using non-batch methods still apply here.
-fn set_all_tiles_with_func<T>(
-    layer_builder: &mut LayerBuilder<T>,
-    mut func: impl FnMut(TilePos) -> Option<T>,
-) where
-    T: TileBundleTrait,
-{
-    let map_size: Vec2 = layer_builder.settings.map_size.into();
-    let chunk_size: Vec2 = layer_builder.settings.chunk_size.into();
-    let map_size_in_tiles = (map_size * chunk_size).as_uvec2();
-    for x in 0..map_size_in_tiles.x {
-        for y in 0..map_size_in_tiles.y {
-            let tile_pos = TilePos(x, y);
-            func(tile_pos).map(|t| layer_builder.set_tile(tile_pos, t).unwrap());
-        }
-    }
-}
-
-fn calculate_transform_from_ldtk_info(
-    location: IVec2,
-    pivot: Vec2,
-    def_size: IVec2,
-    size: IVec2,
-    level_height: i32,
-    layer_id: usize,
-) -> Transform {
-    let pivot_point = Vec2::new(location.x as f32, (level_height - location.y) as f32);
-
-    let adjusted_pivot = Vec2::new(0.5 - pivot.x, pivot.y - 0.5);
-
-    let offset = size.as_vec2() * adjusted_pivot;
-
-    let translation = pivot_point + offset;
-
-    let scale = size.as_vec2() / def_size.as_vec2();
-
-    Transform::from_xyz(translation.x, translation.y, layer_id as f32)
-        .with_scale(Vec3::new(scale.x, scale.y, 1.))
-}
-
-fn calculate_transform_from_entity_instance(
-    entity_instance: &EntityInstance,
-    entity_definition_map: &HashMap<i32, &EntityDefinition>,
-    level_height: i32,
-    layer_id: usize,
-) -> Transform {
-    let entity_definition = entity_definition_map.get(&entity_instance.def_uid).unwrap();
-
-    let location = IVec2::from_slice(entity_instance.px.as_slice());
-
-    let pivot = Vec2::from_slice(entity_instance.pivot.as_slice());
-
-    let def_size = IVec2::new(entity_definition.width, entity_definition.height);
-
-    let size = IVec2::new(entity_instance.width, entity_instance.height);
-
-    calculate_transform_from_ldtk_info(location, pivot, def_size, size, level_height, layer_id)
-}
-
-fn calculate_transform_from_tile_pos(
-    tile_pos: TilePos,
-    tile_size: i32,
-    layer_id: usize,
-) -> Transform {
-    let tile_pos: UVec2 = tile_pos.into();
-    let tile_size = Vec2::splat(tile_size as f32);
-    let translation = tile_size * Vec2::splat(0.5) + tile_size * tile_pos.as_vec2();
-
-    Transform::from_xyz(translation.x, translation.y, layer_id as f32)
 }
